requirements.txt
torch==1.11.0
torchvision==0.12.0
torchmetrics==0.9.2
opencv-python==4.5.1.48
einops==0.4.1
matplotlib==3.2.2
numpy==1.21.6
pandas==1.3.5
wandb==0.12.21
tqdm==4.64.0
Pillow==7.1.2
scikit-learn==1.0.2